+++ packages/expo-router/CHANGELOG.md
@@ -14,6 +14,7 @@
 - [web] Add styled modals and sheets on web with a custom modal stack using vaul ([#37767](https://github.com/expo/expo/pull/37767) by [@hirbod](https://github.com/hirbod))
 - [web] Add transparent modal support and allow modal stacking ([#37856](https://github.com/expo/expo/pull/37856) by [@hirbod](https://github.com/hirbod))
 - [web] Use web modal in `Modal` ([#37732](https://github.com/expo/expo/pull/37732) by [@Ubax](https://github.com/Ubax))
+- Add modal portal ([#37938](https://github.com/expo/expo/pull/37938) by [@Ubax](https://github.com/Ubax))
 
 ### 🐛 Bug fixes